Extracts streaming utilities
